start splitting shadow dom ex into component dirs
