+++ examples/shadow-dom/counter-app/about.jade
