+++ examples/shadow-dom/counter-app/app.jade
