+++ examples/shadow-dom/counter-app/counter.jade
